+++ app/static/lastUpdated.js
@@ -1,3 +1,3 @@
 export default {
-  timestamp: '2025-04-15T16:38:09.210Z'
+  timestamp: '2025-08-14T20:05:26.073Z'
 }